+++ lib/src/fields/checkbox_list_form_field.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../extensions/formx_extension.dart';
+import '../extensions/formx_state.dart';
 import 'widgets/formx_field.dart';
 
 /// A `FormField<List<T>>` that builds a list of [CheckboxListTile].